+++ lib/src/directus_websocket.dart
@@ -83,6 +83,11 @@ class DirectusWebSocket {
           return onDelete(data);
         }
       }
+
+      if (data["event"] == "unsubscribe") {
+        subscriptionDataList
+            .removeWhere((element) => element.uid == data["uid"]);
+      }
     }
   }
 
@@ -114,5 +119,17 @@ class DirectusWebSocket {
     }));
 
     return "refresh token request sent";
+  }
+
+  addSubscription(DirectusWebSocketSubscription subscription) {
+    subscriptionDataList.add(subscription);
+    _channel.sink.add(subscription.toJson());
+  }
+
+  removeSubscription({required String uid}) {
+    _channel.sink.add(jsonEncode({
+      "type": "unsubscribe",
+      "uid": uid,
+    }));
   }
 }