+++ test/directus_websocket_test.dart
@@ -149,5 +149,19 @@ main() {
           ]);
       expect(() => sut.listenSocket(message), throwsException);
     });
+
+    test("Socket receive unsubscribe message", () {
+      final String message =
+          '{"type":"subscription","event":"unsubscribe","data":[{"id":"abc-123","uploaded_by":{"id":"123456"}},{"id":"abc-123","uploaded_by":{"id":"123456"}}],"uid":"itemTest"}';
+      sut.listenSocket(message);
+      expect(sut.subscriptionDataList.length, 0);
+    });
+
+    test("Add subscription", () {
+      sut.addSubscription(DirectusWebSocketSubscription<DirectusItemTest>(
+          uid: "itemTest2", onCreate: onCreate));
+      expect(sut.subscriptionDataList.length, 2);
+      expect(sut.subscriptionDataList.last.uid, "itemTest2");
+    });
   });
 }